task_2/requirements.txt
alembic==1.0.8
amqp==2.4.2
APScheduler==3.5.3
attrs==19.1.0
billiard==3.5.0.5
blinker==1.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
kombu==4.4.0
Mako==1.0.7
MarkupSafe==1.1.1
PyMySQL==0.9.3
pyrsistent==0.14.11
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.1
tzlocal==1.5.1
urllib3==1.24.1
vine==1.2.0
Werkzeug==0.14.1
WTForms==2.2.1
